分离测试
