--- .flake8
@@ -1,6 +1,0 @@
-[flake8]
-max-line-length = 120
-extend-ignore = E203
-exclude =
-    .git
-    __pycache__
