+++ .github/workflows/Test Main Program.yml
@@ -1,0 +1,40 @@
+name: Test Main Program
+
+on:
+  push:
+    branches:
+      - dev
+      - master
+    paths:
+      - '**.py'
+      - '**.yml'
+permissions:
+  contents: read
+
+jobs:
+  test:
+    runs-on: ubuntu-latest
+
+    steps:
+      - uses: actions/checkout@v3
+      - name: Cache PDM dependencies
+        id: check_cache
+        uses: actions/cache@v3
+        with:
+          path: ~/.cache/pdm
+          key: ${{ runner.os }}-pdm-${{ hashFiles('**/pyproject.toml') }}
+      - name: Set up Python 3.11
+        uses: actions/setup-python@v3
+        with:
+          python-version: "3.11"
+      - name: Install PDM
+        run: |
+          pip install pdm
+      - name: Install dependencies
+        if: steps.check_cache.outputs.cache-hit != 'true'
+        run: |
+          python -m pip install --upgrade pip
+          pdm install
+      - name: Test with unittest
+        run: |
+          pdm run python -m unittest tests.unit_test
