+++ .pdm-python
@@ -1,0 +1,1 @@
+/mnt/data/Project/GameYamlSpiderAndGenerator/.venv/bin/python
