+++ pyproject.toml
@@ -48,21 +48,3 @@ name = "gameyamlspiderandgenerator"
 version = "1.7.7"
 description = "A useful tool for generating Furrygameindex yaml files"
 readme = "README.md"
-[tool.ruff]
-select = [
-    "B", # flake8-bugbear
-    "C4", # flake8-comprehensions
-    "E", # pycodestyle - Error
-    "F", # Pyflakes
-    "I", # isort
-    "W", # pycodestyle - Warning
-    "UP", # pyupgrade
-]
-ignore = [
-    "E501", # line-too-long
-    "W191", # tab-indentation
-]
-include = ["**/*.py", "**/*.pyi", "**/pyproject.toml"]
-
-[tool.ruff.pydocstyle]
-convention = "google"
