+++ tests/unit_test.py
@@ -3,11 +3,7 @@ import sys
 import unittest
 from pathlib import Path
 
-from yamlgenerator_hook_search import Search
-from yamlgenerator_hook_validate import Verify
-
 from gameyamlspiderandgenerator.util.config import config
-from gameyamlspiderandgenerator.util.fgi import template_dict
 
 config.load(Path(__file__).parent / "test_config.yaml")
 print(config)
@@ -50,14 +46,6 @@ class SpiderUnitTest(unittest.TestCase):
         from gameyamlspiderandgenerator.util.spider import get_bytes, get_text
         self.assertIsInstance(get_bytes("https://example.com/"), bytes)
         self.assertIsInstance(get_text("https://example.com/"), str)
-
-
-class HookUnitTest(unittest.TestCase):
-    def test_search(self):
-        self.assertIsInstance(Search().setup({**template_dict, 'name': 'dead-space'}), dict)
-
-    def test_verify(self):
-        self.assertIsInstance(Verify().setup(template_dict), dict)
 
 
 if __name__ == "__main__":
